Refactor to register/unregister pattern
